graphql-alt: merge Address with Addressable
## Description

Originally, I was porting `Addressable.asAddress` to GraphQL Beta, when
I realised that there is nothing that an `Address` can do that an
`Addressable` cannot. Based on that observation, I decided to merge the
two types together, getting rid of `Addressable` as a separate concept
compared to `Address` (but retaining the `IAddressable` interface that
is shared among all entities that have an address).

There are some parts of the GraphQL Alpha schema, related to its `Owner`
type, that were going to be attached to `Addressable`, but which now
will go on `Address`:

- `Query.address` will gain a `rootVersion` and an `atCheckpoint`
  parameter.
- `Address` will gain an `asObject` "casting" operation, to try and
  fetch this address as an object.

## Test plan

Existing tests.
