+++ crates/sui-indexer-alt-graphql/src/api/query.rs
@@ -15,7 +15,6 @@ use super::{
     scalars::{digest::Digest, sui_address::SuiAddress, uint53::UInt53},
     types::{
         address::Address,
-        addressable::Addressable,
         checkpoint::Checkpoint,
         epoch::Epoch,
         move_package::{self, CheckpointFilter, MovePackage, PackageKey},
@@ -39,8 +38,7 @@ impl Query {
     /// Look-up an account by its SuiAddress.
     async fn address(&self, ctx: &Context<'_>, address: SuiAddress) -> Result<Address, RpcError> {
         let scope = self.scope(ctx)?;
-        let addressable = Addressable::with_address(scope, address.into());
-        Ok(Address::new(addressable))
+        Ok(Address::with_address(scope, address.into()))
     }
 
     /// First four bytes of the network's genesis checkpoint digest (uniquely identifies the network), hex-encoded.
