+++ crates/sui-indexer-alt-graphql/src/api/types/mod.rs
@@ -2,7 +2,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub(crate) mod address;
-pub(crate) mod addressable;
 pub(crate) mod checkpoint;
 pub(crate) mod epoch;
 pub(crate) mod move_package;
